feat: move literal into core
used by both AST and HIR
