+++ libs/@local/hashql/core/src/lib.rs
@@ -22,6 +22,7 @@ pub mod collection;
 pub mod heap;
 pub mod id;
 pub mod intern;
+pub mod literal;
 pub mod math;
 pub mod module;
 pub mod span;
